+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+#### Added
+- Notice when a symbol cannot be found by light bulb ([#123](https://github.com/buehler/typescript-hero/issues/123))
+
+#### Fixed
+- All possible found declarations are listen in light-blub ([#123](https://github.com/buehler/typescript-hero/issues/123))
 
 ## [0.10.0]
 #### Added